refactor(auth): tighten gateway identity flow

+++ services/user-api/internal/logic/delete_ticket_user_logic.go
@@ -6,11 +6,14 @@ package logic
 import (
 	"context"
 
+	"damai-go/pkg/xerr"
 	"damai-go/services/user-api/internal/svc"
 	"damai-go/services/user-api/internal/types"
 	"damai-go/services/user-rpc/userrpc"
 
 	"github.com/zeromicro/go-zero/core/logx"
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/status"
 )
 
 type DeleteTicketUserLogic struct {
@@ -28,6 +31,22 @@ func NewDeleteTicketUserLogic(ctx context.Context, svcCtx *svc.ServiceContext) *
 }
 
 func (l *DeleteTicketUserLogic) DeleteTicketUser(req *types.DeleteTicketUserReq) (resp *types.BoolResp, err error) {
+	userID, err := requireCurrentUserID(l.ctx)
+	if err != nil {
+		return nil, err
+	}
+
+	ticketUsers, err := l.svcCtx.UserRpc.ListTicketUsers(l.ctx, &userrpc.ListTicketUsersReq{
+		UserId: userID,
+	})
+	if err != nil {
+		return nil, err
+	}
+
+	if !containsTicketUserID(ticketUsers, req.ID) {
+		return nil, status.Error(codes.Unauthenticated, xerr.ErrUnauthorized.Error())
+	}
+
 	rpcResp, err := l.svcCtx.UserRpc.DeleteTicketUser(l.ctx, &userrpc.DeleteTicketUserReq{
 		Id: req.ID,
 	})
@@ -36,4 +55,18 @@ func (l *DeleteTicketUserLogic) DeleteTicketUser(req *types.DeleteTicketUserReq)
 	}
 
 	return mapBoolResp(rpcResp), nil
+}
+
+func containsTicketUserID(resp *userrpc.ListTicketUsersResp, ticketUserID int64) bool {
+	if resp == nil {
+		return false
+	}
+
+	for _, item := range resp.List {
+		if item != nil && item.Id == ticketUserID {
+			return true
+		}
+	}
+
+	return false
 }
